Fix GradientLine's coloring
GradientLine(0, 0, GetScreenWidth(), GetScreenHeight(), CreateColor(255, 0, 0), CreateColor(0, 0, 255)) now makes a line that starts red in the top left, and fades to blue at the lower right - e.g. standard and co drivers now match spheregl

Also adding surface.gradientLine, surface.gradientRectangle and surface.gradientTriangle
The possibly neat thing here is surface.setBlendMode(ALPHA_ONLY), surface.gradient... etc kinds of combinations.

:)


git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@1058 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/editor/MapView.hpp
@@ -26,8 +26,8 @@ public:
 
   BOOL Create(CDocumentWindow* owner, IMapViewHandler* handler, CWnd* parent, sMap* map);
 
-  void SetZoomFactor(int factor);
-  int  GetZoomFactor();
+  void SetZoomFactor(double factor);
+  double  GetZoomFactor();
 
   void TilesetChanged();
 
@@ -132,7 +132,7 @@ private:
   CDIBSection* m_BlitTile;  // for rendering tiles
   CClipboard* m_Clipboard;
 
-  int m_ZoomFactor;
+  double m_ZoomFactor;
   RGB m_ObstructionColor;
   RGB m_HighlightColor;
   RGB m_ZoneColor;
